requirements.txt
einops==0.8.1
ipython==8.12.3
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow>=9.5.0,<11.0.0
s3fs==2024.10.0
timm==0.9.2
torch==2.4.1
tqdm==4.65.2
